Try moving UO2 tests into test_materials
